Ok, The best times stuff looks like it can work. I left it in the compile because I want any problems with it found.
Right now it doesn't do any encryption of the data files because it allows
for you to check them by hand after so many games to see if they are correct.

When it looks stable, I'll add the encryption stuff. Now I need to add
conversion of old best time file stuff and I think that would be it.