Added ticker code to relevant files. It ticks and shows you in the window. Look in tick.c to remove the display of it. Oh, and it saves the time correctly in the best times file and sorts stuff right.
Unfortunately, napms uses usleep which fucks and deferes the sigalarm until
it finishes. THis means that whenever napms gets called, the timer doesn't
work.

So I commented it out of the code and the timer works like it is supposed to.
THe only thing is that I don't know why it was there and whether or not it
should be there. I don't see a difference in game play.