Cleaned up some stuff in Makefile.in ; now the compilation can be observed.
./configure now correctly looks for ncurses in /usr/local, and will
fall back correctly. --with-ncurses=location will also work.

Changed sweep.h.in to have NCURSES_HINT instead of NCURSES. This
complies with the changes to configure.

Changed gpl.c to handle keystrokes differently. 'q' quits the GNU GPL,
instead of exiting the program. Any other key continues; at the end,
any key returns to the game. This eliminates accidental exiting of
the program. The same changes will be made to Help(), along with the
code to stop the timer.