Added the 'n' key for New Game.
Added StartTimer() and StopTimer() to utils.c for a standardized way
of doing that type of shit. Now the menu displays correctly. Make sure
that when you want the timer stopped, wrap whatever it is by those two
functions.

Added the ABORT and RECONF features to game->status. ABORT is implemented
while RECONF needs a better char than 'r' because that is the redraw
character.