Added PrintStats() in drawing.c Now you have a pretty window to see your progress in the game.
One caveat. The stats window is created and destroyed each time it is
updated. I don't know if that is a good idea, all I wanted to do was clear
it before I wrote in it. I couldn't find a function to do that.

But does it really matter anyway?