Fixed the percentage placement in the stats window.
Can't figure out why best times display quits automatically with
a ^A; it's not getting a keystroke, and it should be in the right
input mode.

Took 'got no keystroke' out of play.c; it's pretty obvious that
halfdelay mode works, so may as well not have useless debug messages.