Added more to the manual page.
Starting moving stuff out of defaults.h ; the original intent
of defaults.h was for the user to edit it at compile time, but
using autoconf makes this sort of stupid.

Added "secondary" function searches to configure.in for getopt(),
sprintf() and vsprintf(). These are investigated if and only if
their superior cousins aren't found: getopt_long(), snprintf(),
and vsnprintf().

Added some #if defined stuff to the code around the "lesser"
function calls (see above).