Changed configure.in to not use a debug.log unless explicitly requested; to only use group times if explicitly requested..
Changed sweep.h.in to include the ANSI C macro using the magic
octothrope (#) for mkstr(s) #s. This allows quoting of strings
automatically, which makes using stuff that gets #define'd a
lot easier.

The man page got rebuilt. It does that.

Fixed pbests.c to use the mkstr(s) macro.

Feedback is starting to come in from the field.